Code/requirements.txt
jieba
django
pyahocorasick
requests
pickle
flask